Moved unit tests to its own test file.
